Split widgets module
It was getting ridiculous; time to split it into separate sub-modules
but keep the top-level "widgets" namespace for the monitor app
